RagModule/requirements_rag.txt
langchain
langchain-community
langchain-google-genai
langchain-huggingface
sentence-transformers
faiss-cpu
biopython
python-dotenv
pandas
